+++ src/printer_model_ligo.ml
@@ -459,6 +459,13 @@ let pp_model_internal fmt (model : model) b =
 
     (* effect *)
 
+    | Mfail (Invalid mt) when Utils.is_not_string_nat_int mt.type_ ->
+      let t = mt.type_ in
+      Format.fprintf fmt "if True then fail_%a (%a) else skip;"
+        (* pp_type t *)
+        pp_pretty_type t
+        f mt
+
     | Mfail ft ->
       Format.fprintf fmt "failwith (%a)"
         (pp_fail_type f) ft
@@ -1264,6 +1271,11 @@ let pp_model_internal fmt (model : model) b =
 
     | Mceil x ->
       Format.fprintf fmt "ceil (%a)"
+        f x
+
+    | Mtostring (t, x) ->
+      Format.fprintf fmt "to_string_%a (%a)"
+        pp_pretty_type t
         f x
 
     | Mpack x ->
@@ -2440,6 +2452,48 @@ let pp_model_internal fmt (model : model) b =
 
     | Bfloor    -> Format.fprintf fmt "function floor (const r : %a) : int is block {skip} with r.0 / int(r.1)@\n" pp_type Utils.type_rational
     | Bceil     -> Format.fprintf fmt "function ceil (const r : %a) : int is block {skip} with r.0 / int(r.1) + (if r.0 mod r.1 = 0n then 0n else 1n)@\n" pp_type Utils.type_rational
+    | Btostring t ->
+      Format.fprintf fmt
+        "function to_string_%a (const n : %a) : string is@\n  \
+         block {@\n    \
+         const res : string = \"0\";@\n    \
+         if (n > 0n) then block {@\n      \
+         var tmp : string := \"\";@\n      \
+         var v : nat := n;@\n      \
+         const m : map(nat, string) = (map [@\n        \
+         0n -> \"0\";@\n        \
+         1n -> \"1\";@\n        \
+         2n -> \"2\";@\n        \
+         3n -> \"3\";@\n        \
+         4n -> \"4\";@\n        \
+         5n -> \"5\";@\n        \
+         6n -> \"6\";@\n        \
+         7n -> \"7\";@\n        \
+         8n -> \"8\";@\n        \
+         9n -> \"9\";@\n        \
+         ] : map(nat, string));@\n      \
+         while (v > 0n) block {@\n        \
+         const i : nat = v mod 10n;@\n        \
+         const str : string = get_force(i, m);@\n        \
+         tmp := String.concat(str, tmp);@\n        \
+         v := v / 10n;@\n      \
+         };@\n    \
+         res := tmp;@\n    \
+         } else skip;@\n  \
+         } with res@\n"
+        pp_pretty_type t
+        pp_type t
+
+    | Bfail t ->
+      Format.fprintf fmt
+        "function fail_%a(const v : %a) : unit is@\n  \
+         block {@\n  \
+         const f : (%a -> unit) = [%%Michelson ({| { FAILWITH } |} : %a -> unit)];@\n  \
+         } with f(v)@\n"
+        pp_pretty_type t
+        pp_type t
+        pp_type t pp_type t
+
   in
 
   let pp_api_internal (_env : env) fmt = function
@@ -2447,7 +2501,7 @@ let pp_model_internal fmt (model : model) b =
       Format.fprintf fmt
         "function rat_eq (const lhs : %a; const rhs : %a) : bool is@\n  \
          block {skip} with@\n  \
-         lhs.0 * int(rhs.1) = rhs.0 * int(lhs.1) @\n"
+         lhs.0 * int(rhs.1) = rhs.0 * int(lhs.1)@\n"
         pp_type Utils.type_rational
         pp_type Utils.type_rational
     | RatCmp       ->