+++ src/printer_model_smartpy.ml
@@ -353,6 +353,8 @@ let pp_model fmt (model : model) =
     | Boptget t -> Format.fprintf fmt "getopt on %a" pp_type t
     | Bfloor    -> pp_str fmt "floor"
     | Bceil     -> pp_str fmt "ceil"
+    | Btostring t -> Format.fprintf fmt "to_string on %a" pp_type t
+    | Bfail t     -> Format.fprintf fmt "fail on %a" pp_type t
   in
 
   let pp_api_internal fmt = function
@@ -1211,6 +1213,10 @@ let pp_model fmt (model : model) =
 
       | Mceil x ->
         Format.fprintf fmt "ceil (%a)"
+          f x
+
+      | Mtostring (_, x) ->
+        Format.fprintf fmt "to_string (%a)"
           f x
 
       | Mpack x ->