+++ CHANGES.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+ - `to_string` convert argument to a string (only avalaible for `nat` type for now)
+
+### Changed
+ - `fail` can take any type argument
+
 ## [1.1.0] - 2020-08-05
 ### Added
  - `chain_id` constant